+++ dotflow/core/dotflow.py
@@ -73,3 +73,6 @@ class DotFlow:
             list (List[Any]): Returns a list of assorted objects.
         """
         return [task.current_context.storage for task in self.task.queue]
+
+    def result(self):
+        return self.task.result()